Centred 'to-top' arrow
Had to replace the Font Awesome arrow with a pure CSS one (made of a
rotated square with a border only on two sides). Then I used flexbox to
get it perfectly centred within the circle. This was more effort than
should probably have been expended on a fucking button.